requirements.txt
aiohttp
aiohttp-retry
aiosignal
asgiref
attrs
blinker
certifi
cffi
chardet
charset-normalizer
click
colorama
crispy-bootstrap5
crispy-tailwind
cryptography
defusedxml
dj-database-url
dj-rest-auth
Django
django-allauth
django-cors-headers
django-crispy-forms
django-email
django-environ
django-filter
django-rest-auth
django-tailwind
djangorestframework
djangorestframework-simplejwt
Flask
frozenlist
googletrans
h11
h2
hpack
hstspreload
httpcore
httpx
hyperframe
idna
itsdangerous
Jinja2
Markdown
MarkupSafe
multidict
oauthlib
pillow
psycopg2
psycopg2-binary
pycparser
PyJWT
PyPDF2
python3-openid
pytz
reportlab
requests
requests-oauthlib
rfc3986
six
sniffio
sqlparse
twilio
typing_extensions
tzdata
urllib3
Werkzeug
yarl
